\name{html_clean}
\alias{html_clean}
\title{Convert to an HTML document}
\usage{
  html_clean(number_sections = FALSE, fig_width = 6,
    fig_height = 6, fig_retina = if (!fig_caption) 2,
    fig_caption = TRUE, smart = TRUE,
    self_contained = TRUE, highlight = "pygments",
    mathjax = "default", css = NULL, includes = NULL,
    keep_md = FALSE, lib_dir = NULL, pandoc_args = NULL,
    ...)
}
\arguments{
  \item{number_sections}{\code{TRUE} to number section
  headings}

  \item{fig_width}{Default width (in inches) for figures}

  \item{fig_height}{Default width (in inches) for figures}

  \item{fig_retina}{Scaling to perform for retina displays
  (defaults to 2 when \code{fig_caption} is \code{FALSE},
  which currently works for all widely used retina
  displays). Set to \code{NULL} to prevent retina scaling.
  Note that this will always be \code{NULL} when
  \code{keep_md} is specified (this is because
  \code{fig_retina} relies on outputting HTML directly into
  the markdown document).}

  \item{fig_caption}{\code{TRUE} to render figures with
  captions}

  \item{smart}{Produce typographically correct output,
  converting straight quotes to curly quotes, --- to
  em-dashes, -- to en-dashes, and ... to ellipses.}

  \item{self_contained}{Produce a standalone HTML file with
  no external dependencies, using data: URIs to incorporate
  the contents of linked scripts, stylesheets, images, and
  videos. Note that even for self contained documents
  MathJax is still loaded externally (this is necessary
  because of it's size).}

  \item{highlight}{Syntax highlighting style. Supported
  styles include "default", "tango", "pygments", "kate",
  "monochrome", "espresso", "zenburn", "haddock", and
  "textmate". Pass \code{NULL} to prevent syntax
  highlighting.}

  \item{mathjax}{Include mathjax. The "default" option uses
  an https URL from the official MathJax CDN. The "local"
  option uses a local version of MathJax (which is copied
  into the output directory). You can pass an alternate URL
  or pass \code{NULL} to exclude MathJax entirely.}

  \item{css}{One or more css files to include}

  \item{includes}{Named list of additional content to
  include within the document (typically created using the
  \code{\link{includes}} function).}

  \item{keep_md}{Keep the markdown file generated by
  knitting.}

  \item{lib_dir}{Directory to copy dependent HTML libraries
  (e.g. jquery, bootstrap, etc.) into. By default this will
  be the name of the document with \code{_files} appended
  to it.}

  \item{pandoc_args}{Additional command line options to
  pass to pandoc}

  \item{...}{Additional function arguments to pass to the
  base R Markdown HTML output formatter}
}
\value{
  R Markdown output format to pass to \code{\link{render}}
}
\description{
  Format for converting from R Markdown to an HTML
  document.
}
\examples{
\dontrun{

library(rmdformats)

render("input.Rmd", html_clean())

}
}

