% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilltabs.R
\name{knit_print.pilltabs}
\alias{knit_print.pilltabs}
\title{knitr printing function for pilltabs}
\usage{
\method{knit_print}{pilltabs}(res, ...)
}
\arguments{
\item{res}{data to be printed, generated by \code{\link{pilltabs}}}

\item{...}{arguments passed to other methods}
}
\description{
Not to be used directly
}

