% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockdown.R
\name{lockdown}
\alias{lockdown}
\title{Convert to an HTML document}
\usage{
lockdown(...)
}
\arguments{
\item{...}{Additional function arguments passed to R Markdown \code{\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for converting from R Markdown to an HTML document.
}
\details{
TODO
}
