% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilltabs.R
\name{print.pilltabs}
\alias{print.pilltabs}
\alias{knit_print.pilltabs}
\title{Printing functions for pilltabs}
\usage{
\method{print}{pilltabs}(x, ...)

\method{knit_print}{pilltabs}(x, ...)
}
\arguments{
\item{x}{data to be printed, generated by \code{\link{pilltabs}}}

\item{...}{arguments passed to other methods}
}
\description{
Not to be used directly
}
