% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readthedown.R
\name{readthedown}
\alias{readthedown}
\title{Convert to an HTML document}
\usage{
readthedown(fig_width = 8, fig_height = 5, fig_caption = TRUE,
  highlight = "kate", pandoc_args = NULL, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", "haddock", and "textmate". Pass \code{NULL} to prevent syntax
highlighting.}

\item{pandoc_args}{arguments passed to the pandoc_args argument of rmarkdown \code{\link{html_document}}}

\item{...}{Additional function arguments passed to R Markdown \code{\link{html_document}}}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to an HTML document.
}

