% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robobook.R
\name{robobook}
\alias{robobook}
\title{Convert to an HTML document}
\usage{
robobook(
  fig_width = 8,
  fig_height = 5,
  fig_caption = TRUE,
  highlight = "kate",
  lightbox = FALSE,
  thumbnails = FALSE,
  gallery = FALSE,
  toc_depth = 2,
  embed_fonts = TRUE,
  use_bookdown = FALSE,
  pandoc_args = NULL,
  md_extensions = NULL,
  mathjax = "rmdformats",
  ...
)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", "haddock", and "textmate". Pass \code{NULL} to prevent syntax
highlighting.}

\item{lightbox}{if TRUE, add lightbox effect to content images}

\item{thumbnails}{if TRUE display content images as thumbnails}

\item{gallery}{if TRUE and lightbox is TRUE, add a gallery navigation between images in lightbox display}

\item{toc_depth}{adjust table of contents depth}

\item{embed_fonts}{if TRUE, use local files for fonts used in the template. This leads to bigger files but ensures that these fonts are available. If FALSE they are downloaded from Google Web Fonts.}

\item{use_bookdown}{if TRUE, uses \code{\link[bookdown]{html_document2}} instead of \code{\link[rmarkdown]{html_document}}, thus providing numbered sections and cross references}

\item{pandoc_args}{arguments passed to the pandoc_args argument of rmarkdown \code{\link[rmarkdown]{html_document}}}

\item{md_extensions}{arguments passed to the md_extensions argument of rmarkdown \code{\link[rmarkdown]{html_document}}}

\item{mathjax}{set to NULL to disable Mathjax insertion}

\item{...}{Additional function arguments passed to R Markdown \code{\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for converting from R Markdown to an HTML document.
}
\details{
CSS adapted from the bookdown project, with Roboto family fonts.
}
